clicks

+++ core/src/com/trance/empire/modules/replay/model/Click.java
@@ -3,35 +3,35 @@ package com.trance.empire.modules.replay.model;
 public class Click {
 	
 	/**
-	 * 相对于开始时间+
+	 * 倒计时
 	 */
 	private int t;
 	
-	private float x;
+	private int x;
 	
-	private float y;
+	private int y;
 
-	public int getT() {
-		return t;
-	}
+    public int getT() {
+        return t;
+    }
 
-	public void setT(int t) {
-		this.t = t;
-	}
+    public void setT(int t) {
+        this.t = t;
+    }
 
-	public float getX() {
-		return x;
-	}
+    public int getX() {
+        return x;
+    }
 
-	public void setX(float x) {
-		this.x = x;
-	}
+    public void setX(int x) {
+        this.x = x;
+    }
 
-	public float getY() {
-		return y;
-	}
+    public int getY() {
+        return y;
+    }
 
-	public void setY(float y) {
-		this.y = y;
-	}
+    public void setY(int y) {
+        this.y = y;
+    }
 }
